Save Scene 12:
Professor saved the day. It was the nullException error that I kept ignoring that was the problem, LOL. I didn't think it was the problem since the error showed up prior to the disappearance of the cards (and I was gonna fix it later) but I was wrong. Basically I didn't assign the Highscore, Gameover, and Roundscore Texts in the inspector.