P2 Save Scene 3:
Layout.xml is finally good, kinda. The only problem now is that when in game mode, the pyramid does not look stacked behind each other, but in the scene they are. I think this is a problem with Layout.cs so i'm going to tinker with that. Also i realize that I rewrote over the original Prospector, which I do not know if I was supposed to or not. I will ask in class.