P2 Save Scene 5:
Added a new function "AddUpToThirteen" and removed/commented out the function "AdjacentRank". The cards are somewhat on the Tableau however they are not in the right order, and the error message is bringing me to Prospector (probably because I copied and pasted most of my elements from there). Going to try to fix the layout and then find out how to click two cards