P2 - Save Scene 9:
I don't have enough time to figure out how to implement clicking two cards on the board. I know what I have to do in concept: Store the card that the player first picks with the card the player picks right after. If it's sum equals 13, move to discard; if it is 13, move to discard. If none of these occur, reset the selection.
So I'm leaving it at this, the player can move cards to the discard pile but only comparing it with the target card. Or if it is the king.